fix: move takeDataForm to utils
